Integrate convolution into cffi
