+++ rosetta/src/main/cpp/app/cffi_run.py
@@ -1,4 +1,4 @@
-from _fullyconnected import lib, ffi
+from _qbart import lib, ffi
 import numpy as np
 import math
 import random
